Minor changes in rides
