+++ app/domain/services/ride.py
@@ -1,7 +1,8 @@
 from fastapi import HTTPException
 
 from app.domain.models import User, Schedule
-from app.infrastructure.repository.ride import RideData, RideRepository
+from app.domain.types import RideData
+from app.infrastructure.repository.ride import RideRepository
 from app.infrastructure.repository.travels import ScheduleRepository
 from app.infrastructure.repository.user import UserRepository
 from app.presentation.schemes import RideRequest
