+++ app/domain/types.py
@@ -14,3 +14,10 @@ class LocationData:
             'latitude': self.latitude,
             'longitude': self.longitude,
         }
+
+
+@dataclasses.dataclass
+class RideData:
+    location: str
+    latitude: str
+    longitude: str
