+++ app/infrastructure/repository/tracking.py
@@ -1,0 +1,7 @@
+from app.core.dependencies.depends.database.redis import CacheSession
+
+
+class TrackingRepository:
+    @staticmethod
+    async def record(cache: CacheSession):
+        pass
